requirements.txt
pytesseract
Pillow
requests
pandas
google-cloud-aiplatform
google-generativeai
pdf2image
python-poppler
